services/asr_service/requirements.txt
fastapi
uvicorn
openai-whisper
torch
torchaudio
pydantic
